requirements.txt
Flask==3.0.0
pytesseract==0.3.10
Pillow>=10.4.0
pytz==2024.1
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
